Create import_qr_code_document_creator.py
Allows the user to select a file that contains some pngs. The pngs are then placed into a table inside a word document. They are spaced out so that the paper they are printed on is used as a label. 

This is intended to be used to print on the 3x5 AVERY label printer paper.