Update import_qr_code_document_creator.py
Fixed the algorithm that creates the number of rows for the table. 

It was previously using a floor divider and adding 1. This caused a situation where if the result was evenly divisible by 3 then it would unintentionally round up the number of rows needed by 1 whole integer. 

The solution was to divide the number normally by three then use the .ceil method in math to round up if any remainder was present. This always produces the correct number of rows.